Converted old RealSeq code to newer FitnessFunc paradigm

+++ src/main/scala/ml/realSeqFunction/Mutation.scala
@@ -3,25 +3,25 @@ package com.RealSeqFunction
 import com.util.Entropy.rand
 
 trait RSMutate {
-    def apply(dna: Seq[Float], min: Float, max: Float): Seq[Float]
+    def apply(dna: Seq[Double], min: Double, max: Double): Seq[Double]
 }
 
-class GaussMutate(sdv: Float) extends RSMutate{
-    def apply(dna: Seq[Float], min: Float, max: Float): Seq[Float] = {
-        def m(v: Float): Float = {
-            val tmp = v + ((rand nextGaussian).toFloat * sdv)
-            Math.min(max, Math.max(tmp, min)).toFloat
+class GaussMutate(sdv: Double) extends RSMutate{
+    def apply(dna: Seq[Double], min: Double, max: Double): Seq[Double] = {
+        def m(v: Double): Double = {
+            val tmp = v + ((rand nextGaussian).toDouble * sdv)
+            Math.min(max, Math.max(tmp, min)).toDouble
         }
         dna.map(m(_))
     }
 }
 
-class SelectiveMutate(mutateChance: Float, sdv: Float) extends RSMutate{
-    def apply(dna: Seq[Float], min: Float, max: Float): Seq[Float] = {
-        def m(v: Float): Float = {
-            if(rand.nextFloat < mutateChance){
-                val tmp = v + ((rand nextGaussian).toFloat * sdv)
-                Math.min(max, Math.max(tmp, min)).toFloat
+class SelectiveMutate(mutateChance: Double, sdv: Double) extends RSMutate{
+    def apply(dna: Seq[Double], min: Double, max: Double): Seq[Double] = {
+        def m(v: Double): Double = {
+            if(rand.nextDouble < mutateChance){
+                val tmp = v + ((rand nextGaussian).toDouble * sdv)
+                Math.min(max, Math.max(tmp, min)).toDouble
             } else {
                 v
             }
